modified:   UNITE_Main.X/SatelliteMode.c 	modified:   UNITE_Main.X/SatelliteMode.h 	modified:   UNITE_Main.X/adc1.c 	modified:   UNITE_Main.X/adc1.h 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/SatelliteMode.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/adc1.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/main.o.d 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/mcc.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/mcc.o.d 	deleted:    UNITE_Main.X/dist/XC16_24FJ256GA110/production/BackupofMCC.production.elf 	deleted:    UNITE_Main.X/dist/XC16_24FJ256GA110/production/BackupofMCC.production.hex 	deleted:    UNITE_Main.X/dist/XC16_24FJ256GA110/production/BackupofMCC.production.map 	modified:   UNITE_Main.X/dist/XC16_24FJ256GA110/production/memoryfile.xml 	modified:   UNITE_Main.X/mcc_generated_files/mcc.c 	modified:   UNITE_Main.X/mcc_generated_files/mcc.h 	modified:   UNITE_Main.X/nbproject/Makefile-XC16_24FJ256GA110.mk 	modified:   UNITE_Main.X/nbproject/configurations.xml 	modified:   UNITE_Main.X/nbproject/private/configurations.xml 	UNITE_Main.X/dist/XC16_24FJ256GA110/production/UNITE_Main.X.production.elf 	UNITE_Main.X/dist/XC16_24FJ256GA110/production/UNITE_Main.X.production.hex 	UNITE_Main.X/dist/XC16_24FJ256GA110/production/UNITE_Main.X.production.map 	UNITE_Main.X/mcc_generated_files/spi1.c 	UNITE_Main.X/mcc_generated_files/spi1.h 	UNITE_Main.X/mcc_generated_files/spi2.c 	UNITE_Main.X/mcc_generated_files/spi2.h

+++ UNITE_Main.X/SatelliteMode.h
@@ -37,13 +37,13 @@ typedef struct {
 
 
 void Satellite_Initialize(void);
-void GetTempData(int *buffer, int bufferSize);
+void GetTempData(uint8_t *buffer, int bufferSize);
 void GetGPSData(int *buffer, int bufferSize);
 void GetMagnetometerData(int *buffer, int bufferSize);
 void GetProbeData(int *buffer, int bufferSize);
 void SaveData(int *package, int packageSize);
 void PackageData(int *package, int stringLength, int *temps, int *gps, int *mags, int *densities);
-void SendData(int *dataString, int stringLength);
+void SendData(uint8_t *dataString, int stringLength);
 
 UNITEMode UpdateMode(void);
 unsigned int DelayForMode(void);
